// src/app/models/premio.ts
export interface Premio{
  id:number;
  titulo: string;
  descricao: string;
  publicado: string ;
  validade: string ;
  pontos: number ;
  nivel: number ;
  quantidade: number ;
  valido: boolean;
  vencido: boolean;
  utilizado: number;
  image: string;
  nivelnome: string;
  nivelnome_1:string
}